procfs: Avoid costly allocations from buffered reads
procfs is one of the hottest parts of the below codebase, accounting for
roughly a quarter of its overall work. It currently uses BufReader for
reading the majority of files, but this is highly inefficient for small
files:

1. BufReader allocates a buffer in memory, which is typically larger
   than the size of the small file. For small files, this buffer
   allocation can be more costly in terms of time and memory than simply
   reading the file directly without buffering.
2. BufReader is designed to reduce the number of syscalls, but in this
   case we basically only do one (a single read()).

We do the vast majority of reads synchronously (except for cmdline due
to priority inversion avoidance). This allows us to use a shared buffer
for all reads, avoiding paying the cost each time for buffer
instantiation. In some cases where we don't use it as a string at all,
it avoids any allocation at all.

This reduces CPU usage by about 5% on my laptop with a few cgroups. With
more cgroups, as in production, it will be even more.

As an extra bonus, this even reduces the amount of code required.

We do not use std::fs::read_to_string (or the Read trait) because that
unfortunately does this, which is extremely unhelpfyl on procfs with no
inode st_size:

    let size = file.metadata().map(|m| m.len() as usize).ok();
    let mut string = String::new();
    string.try_reserve_exact(size.unwrap_or(0))?;