below: cmdline: Use AtomicPtr for mmap_sem priority inversion avoidance
We use a background thread when reading /proc/pid/cmdline in order to
avoid priority inversions when mmap_sem is held for a long time or is
deadlocked.

We previously used an `mpsc::channel` to communicate between the main
thread and the background thread. While effective, this approach incurs
non-trivial synchronisation overhead due to the underlying locking
mechanisms of the channel. By comparison, this new approach is
essentially free in terms of synchronisation on x86_64.

This saves about 0.5-1% on-CPU time per iteration.