read_pid_cmdline: Use atomic pointer to avoid MPSC channel overhead
We use a background thread when reading /proc/pid/cmdline in order to
avoid priority inversions when mmap_sem is held for a long time or is
deadlocked.

We previously used an `mpsc::channel` to communicate between the main
thread and the background thread. While effective, this approach incurs
non-trivial synchronisation overhead due to the underlying locking
mechanisms of the channel. By comparison, this new approach is
essentially free in terms of synchronisation on x86_64.

In my tests this results in a reduction of around 1% in CPU instructions
per iteration.