uptime: Reduce syscall overhead using clock_gettime(CLOCK_BOOTTIME)
Reading /proc/uptime is very expensive in terms of syscalls and procfs
synchronisation (to be honest I was surprised at how high it is). By
using clock_gettime(), we can directly access the system clock with a
single syscall, which is significantly cheaper in terms of execution
cost.

This saves about 3-4% CPU(!) each iteration in my tests.