below: cmdline: Avoid perverse std::io::Read::read_to_end heuristics
Summary:
Reading a pid's command line is one of the most common operations we
perform, and it shows up as a decent chunk of our on-CPU time for each
collection interval.

As a virtual filesystem, /proc does not set st_size on most files, and
/proc/pid/cmdline is no exception:

    % stat --printf=%s /proc/self/cmdline
    0

std::io::Read's `read_to_end` does not interact particularly well with
this, and procfs in general. In particular, it does a whole bunch of
probing, checks, heuristics, and reallocations that are unhelpful on
procfs.

We can drastically reduce this for the nominal case by instead reading
directly into a &[u8] buffer.

This saves about 0.5-1% on-CPU time per iteration in my tests.

Reviewed By: dschatzberg

Differential Revision: D58083967

fbshipit-source-id: 6be6588372a32305e62752ef63fccd79a6ba30d4