below: procfs: Amortise file buffer allocation
Summary:
procfs is one of the hottest parts of the below codebase, accounting for
roughly a quarter of its overall work. It currently uses BufReader for
reading the majority of files, but this is highly inefficient for small
files:

1. BufReader allocates a buffer in memory, which is typically larger
   than the size of such small procfs files. For small files, this
   buffer allocation can be more costly in terms of time and memory than
   simply reading the file directly without buffering.
2. BufReader is designed to reduce the number of syscalls, but in this
   case we basically only do one (a single read()).

We do the vast majority of reads synchronously (except for cmdline due
to priority inversion avoidance). This allows us to use a shared buffer
for all reads, avoiding paying the cost each time for buffer
instantiation. In some cases where we don't use it as a string at all,
it avoids any allocation at all.

As an extra bonus, this even reduces the amount of code required.

We do not use std::fs::read_to_string (or the Read trait) because that
unfortunately does this, which is extremely unhelpfyl on procfs with no
inode st_size:

    let size = file.metadata().map(|m| m.len() as usize).ok();
    let mut string = String::new();
    string.try_reserve_exact(size.unwrap_or(0))?;

This saves about 2% on-CPU time per iteration in my rudimentary laptop
tests with a few pids. With more pids, as it is in production, it
will be even more.

Reviewed By: brianc118

Differential Revision: D58083972

fbshipit-source-id: 7ae233e83232f866734230d6b4029cce71a501e8