below: uptime: Reduce syscall overhead with clock_gettime(CLOCK_BOOTTIME)
Summary:
Reading /proc/uptime is very expensive in terms of syscalls and procfs
synchronisation (to be honest I was surprised at how high it is). By
using clock_gettime(), we can directly access the system clock with a
single syscall, which is significantly cheaper in terms of execution
cost.

This saves about 3-4% CPU(!) each iteration in my tests.

Reviewed By: dschatzberg

Differential Revision: D58083973

fbshipit-source-id: f97b70cc9bb788dc31672daca42d850291855277